moved functions to semantically more correct headers
